Revert the UTC time change. It fails with FPC 3.2.2.
This reverts commit 657422dc26f0878ede5af0df3dc730fb54e12034.

+++ ide/inspectchksumchangeddlg.pas
@@ -73,7 +73,7 @@ type
   TICCFile = class(TComponent)
   public
     Filename: string;
-    Age: int64;
+    Age: integer;
     OwnerNames: TStringList;
     constructor Create(AOwner: TComponent); override;
     destructor Destroy; override;
